+++ src/asphalt/templating/api.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 
 from abc import ABCMeta, abstractmethod
-
-from asphalt.core.context import Context
+from typing import Any
 
 
 class TemplateRenderer(metaclass=ABCMeta):
@@ -11,7 +10,7 @@ class TemplateRenderer(metaclass=ABCMeta):
     __slots__ = ()
 
     @abstractmethod
-    def render(self, template: str, **vars) -> str:
+    def render(self, template: str, **vars: Any) -> str:
         """
         Render the named template.
 
@@ -22,7 +21,7 @@ class TemplateRenderer(metaclass=ABCMeta):
         """
 
     @abstractmethod
-    def render_string(self, source: str, **vars) -> str:
+    def render_string(self, source: str, **vars: Any) -> str:
         """
         Render the template contained in the given string.
 
@@ -30,52 +29,3 @@ class TemplateRenderer(metaclass=ABCMeta):
         :param vars: extra variables made available to the template
         :return: the rendered results
         """
-
-
-class TemplateRendererProxy(TemplateRenderer):
-    """
-    Context-bound template renderer proxy.
-
-    Adds the bound context to the variables passed to :meth:`~TemplateRenderer.render`
-    as ``ctx``, unless a variable by that name was explicitly passed.
-
-    Any variables and methods provided by the underlying renderer can be directly
-    accessed through this proxy object.
-    """
-
-    __slots__ = "_ctx", "_renderer"
-
-    def __init__(self, ctx: Context, renderer: TemplateRenderer) -> None:
-        self._ctx = ctx
-        self._renderer = renderer
-
-    def __getattr__(self, name):
-        return getattr(self._renderer, name)
-
-    def render(self, template: str, **vars) -> str:
-        """
-        Render the named template.
-
-        The current context will be available to the template as the ``ctx`` variable.
-
-        :param template: name of the template file
-        :param vars: extra template variables
-        :return: the rendered results
-
-        """
-        vars.setdefault("ctx", self._ctx)
-        return self._renderer.render(template, **vars)
-
-    def render_string(self, source: str, **vars) -> str:
-        """
-        Render the template contained in the given string.
-
-        The current context will be available to the template as the ``ctx`` variable.
-
-        :param source: content of the template to render
-        :param vars: extra variables made available to the template
-        :return: the rendered results
-
-        """
-        vars.setdefault("ctx", self._ctx)
-        return self._renderer.render_string(source, **vars)